[cellEditor] extracted a cell code editor from code mirror
